add blueprint and update routes
